+++ .github/workflows/coverage.yml
@@ -10,7 +10,7 @@ jobs:
       - name: Install cargo-llvm-cov
         uses: taiki-e/install-action@cargo-llvm-cov
       - name: Generate code coverage
-        run: cargo +nightly llvm-cov --all-features --doctests --workspace --lcov --output-path lcov.info
+        run: cargo +nightly llvm-cov --all-features --doctests --workspace --ignore-filename-regex "src/docs/.*" --lcov --output-path lcov.info
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
